cleanup and reorganize extras captions
