+++ cloud-minecraft-extras/src/main/java/org/incendo/cloud/minecraft/extras/caption/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * minecraft-extras caption utilities.
+ */
+package org.incendo.cloud.minecraft.extras.caption;
